cognate updates
the cognito identity id is not the key in the user database.
the key in the email database is still the custom provider id an
cognito is used to map email to identity id.
user/me/get now returns the actual email - the identity id is filled in
automatically by api gateway.